Move exceptions to the client package
