+++ mirrulations-work-server/src/mirrserver/get_client_id_validator.py
@@ -1,4 +1,4 @@
-from mirrserver.exceptions import InvalidClientIDException
+from mirrclient.exceptions import InvalidClientIDException
 
 
 class GetClientIDValidator():
