+++ mirrulations-work-server/src/mirrserver/put_results_validator.py
@@ -1,6 +1,6 @@
-from mirrserver.exceptions import InvalidResultsException
-from mirrserver.exceptions import InvalidClientIDException
-from mirrserver.exceptions import MissingClientIDException
+from mirrclient.exceptions import InvalidResultsException
+from mirrclient.exceptions import InvalidClientIDException
+from mirrclient.exceptions import MissingClientIDException
 
 
 class PutResultsValidator():
